GH-113528: Deoptimise `pathlib._abc.PurePathBase.parent[s]`
Replace use of `_from_parsed_parts()` with `with_segments()`, and move
assignments to `_drv`, `_root`, _tail_cached` and `_str` slots into
`PurePath`.
